@@ -1,28 +1,40 @@
 {
-  "name": "merlinideasbot",
+  "name": "meros",
   "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
+  "private": true,
+  "license": "MIT",
+  "type": "module",
+  "main": "build/main.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "keywords": [],
-  "author": "@Ekpy",
-  "license": "ISC",
-  "devDependencies": {
-    "eslint": "^8.23.0"
+    "build": "tsc",
+    "dev": "tsx src/main.ts",
+    "start": "node build/main.js",
+    "watch": "nodemon --exec npm run dev --watch src --ext ts"
   },
   "dependencies": {
-    "@discordjs/opus": "github:discordjs/opus",
-    "@discordjs/rest": "^1.7.1",
-    "@discordjs/voice": "^0.16.0",
-    "axios": "^0.27.2",
-    "discord.js": "^14.11.0",
-    "gtts": "^0.2.1",
-    "https": "^1.0.0",
-    "jquery": "^3.6.1",
-    "jsdom": "^20.0.0",
-    "libsodium-wrappers": "^0.7.10",
-    "string-similarity": "^4.0.4"
+    "@discordjs/voice": "^0.19.0",
+    "@discordx/importer": "^1.3.3",
+    "@discordx/pagination": "^4.4.0",
+    "@snazzah/davey": "^0.1.8",
+    "better-sqlite3": "^12.5.0",
+    "discord.js": "^14.22.1",
+    "discordx": "^11.13.2",
+    "dotenv": "^17.2.3",
+    "gtts": "^0.2.1"
+  },
+  "devDependencies": {
+    "@types/better-sqlite3": "^7.6.13",
+    "@types/node": "^24.10.1",
+    "@typescript-eslint/parser": "^5.62.0",
+    "eslint": "^8.57.1",
+    "nodemon": "^3.1.10",
+    "prettier": "^3.7.3",
+    "prettier-eslint": "^16.4.2",
+    "tsx": "^4.20.5",
+    "typescript": "^4.9.5"
+  },
+  "engines": {
+    "node": ">=16.0.0",
+    "npm": ">=7.0.0"
   }
 }
