@@ -1,8 +1,12 @@
-const { SlashCommandBuilder, codeBlock } = require('discord.js');
+import {
+  ApplicationCommandOptionType,
+  codeBlock,
+  EmbedBuilder,
+  type CommandInteraction,
+} from 'discord.js';
+import { Discord, Slash, SlashChoice, SlashOption } from 'discordx';
 
-
-const chair =
-`
+const chair = `
 [0;31m   __________.
 [0;31m  /_/-----[0;31m/_/|[0;33m   __
 [0;31m  ( ([0;30m ' ' [0;31m( (|[0;33m /[0;37m'--'[0;33m\\
@@ -14,11 +18,10 @@ const chair =
 [0;31m'        '    [0;30m/______\\ 
 `;
 
-module.exports = {
-	data: new SlashCommandBuilder()
-		.setName('chair')
-		.setDescription('Gives nice chair'),
-	async execute(interaction) {
-		await interaction.reply(codeBlock('ansi', chair));
-	},
-};
+@Discord()
+export class ChairCommand {
+  @Slash({ description: 'Get a nice chair' })
+  async chair(interaction: CommandInteraction): Promise<void> {
+    await interaction.reply(codeBlock('ansi', chair));
+  }
+}
